Fix client-safe username validation imports
